+++ apps/public/src/app/(map)/locatie/kies/page.tsx
@@ -1,48 +1,9 @@
-'use client'
+import type { Metadata } from 'next'
 
-import { Button } from '@amsterdam/design-system-react'
-import dynamic from 'next/dynamic'
-import { useState } from 'react'
+import { KiesLocatie } from './KiesLocatie'
 
-import type { Coordinates } from 'apps/public/src/types'
-
-import { AssetList } from './_components/AssetList/AssetList'
-import { SideBar } from './_components/SideBar/SideBar'
-import styles from './page.module.css'
-
-const Map = dynamic(() => import('./_components/Map').then((module) => module.Map), {
-  loading: () => <p>Loading...</p>, // TODO: improve loading state
-  ssr: false,
-})
-
-const KiesLocatie = () => {
-  const [coordinates, setCoordinates] = useState<Coordinates>()
-  const [showAssetList, setShowAssetList] = useState(false)
-
-  const handleAssetListToggle = () => {
-    setShowAssetList((prevState) => !prevState)
-  }
-
-  return (
-    <div className={styles.grid}>
-      <SideBar coordinates={coordinates} setCoordinates={setCoordinates} />
-      <div className={`${styles.assetList} ${showAssetList && styles.showAssetList}`}>
-        <AssetList />
-        <Button form="address" type="submit" className={styles.hideButtonMobile}>
-          Bevestigen
-        </Button>
-      </div>
-      <div className={styles.map}>
-        <Map coordinates={coordinates} setCoordinates={setCoordinates} showAssetList={showAssetList} />
-        <Button form="address" type="submit" className={styles.submitbutton}>
-          Bevestigen
-        </Button>
-        <Button variant="secondary" onClick={handleAssetListToggle} className={styles.toggleButton}>
-          {showAssetList ? 'Kaart' : 'Lijst'}
-        </Button>
-      </div>
-    </div>
-  )
+export const metadata: Metadata = {
+  title: 'Stap 1 van 4 - Beschrijf uw melding - Gemeente Amsterdam',
 }
 
-export default KiesLocatie
+export default async () => <KiesLocatie />
